backup_twitter/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
attrs==17.4.0
docopt==0.6.2
ply==3.10                 # via pyhcl
pyhcl==0.3.10
